Feature/clp more codecov (#75)
POC to evaluate codecov.io UI, and to see work involved in pushing code coverage as close to 100% as possible just for x/clp. I got this code to 96.61%. The remaining 3 lines of code would have required a custom mocked codec that only failed on MarshalJSONIndent in order to get coverage.