Change voting power calculation
Use delegated token amounts rather than bonded token amounts to
calculate voting power. This significantly reduces the complexity of the
calculations. This is especially true for redelegations which have a
number of cases when using bonded tokens. The real world difference in
calculated voting power is negligible since most tokens are delegated to
validators that are bonded.